Moved dayOfWeek Date method extension to it's own file
