+++ UOKiosk/WhatIsOpen/WhatIsOpenViewModel.swift
@@ -9,16 +9,6 @@ import SwiftUI
 import Combine
 import OrderedCollections
 
-extension Date {
-    // https://stackoverflow.com/questions/25533147/get-day-of-week-using-nsdate
-    func dayOfWeek() -> String? {
-        let dateFormatter = DateFormatter()
-        dateFormatter.dateFormat = "EEEE"
-        return dateFormatter.string(from: self).capitalized
-        // or use capitalized(with: locale) if you want
-    }
-}
-
 class WhatIsOpenViewModel: ObservableObject {
     private let whatIsOpenRepository: WhatIsOpenRepository
